College Project Part4: Implemented the Email sending feature

+++ src/components/college/database/buildinghttp.ts
@@ -5,13 +5,9 @@ const DB_URL = "https://react-native-course-5d92a-default-rtdb.firebaseio.com";
 const DB_NAME = "COLLEGEAPP";
 
 export const storeBuildingData = async (buildingData: BuildingDetailsDTO) => {
-    try {
-        const response = await axios.post(`${DB_URL}/${DB_NAME}/building.json`, buildingData);
-        const id = response.data.name;
-        return id;
-    } catch (error) {
-        console.log("Unable to insert data", error);
-    }
+    const response = await axios.post(`${DB_URL}/${DB_NAME}/building.json`, buildingData);
+    const id = response.data.name;
+    return id;
 }
 
 export const fetchBuildingData = async () => {
@@ -34,14 +30,17 @@ export const fetchBuildingData = async () => {
 }
 
 export const fetchBuildingDataById = async (buildingId: string) => {
+    if(!buildingId) throw new Error("Building id is empty");
     const response = await axios.get(`${DB_URL}/${DB_NAME}/building/${buildingId}.json`);
     return response.data;
 }
 
 export const updateBuildingData = async (id: string, newbuilding: BuildingDetailsDTO) => {
+    if(!id) throw new Error("Id is empty");
     return await axios.put(`${DB_URL}/${DB_NAME}/building/${id}.json`, newbuilding);
 }
 
 export const deleteBuildingData = async (id: string) => {
+    if(!id) throw new Error("Id is empty");
     return await axios.delete(`${DB_URL}/${DB_NAME}/building/${id}.json`);
 }
